+++ test/bigtest/interactors/locations.js
@@ -1,6 +1,7 @@
 import {
   interactor,
   isPresent,
+  collection,
 } from '@bigtest/interactor';
 
 import SelectInteractor from '@folio/stripes-components/lib/Select/tests/interactor';
@@ -17,6 +18,7 @@ import MultiColumnListInteractor from '@folio/stripes-components/lib/MultiColumn
   institutionSelect = new SelectInteractor('[data-test-institution-select]');
   campusSelect = new SelectInteractor('[data-test-campus-select]');
   librarySelect = new SelectInteractor('[data-test-library-select]');
+  librarySelectOptions = collection('[data-test-library-select] select option')
 }
 
 export default Locations;